+++ test/Test/TestDiffTime.hs
@@ -4,9 +4,9 @@ import Test.Hspec
 import Debug.Trace
 import Data.Time.Clock
 import Control.Concurrent
-import Data.Time.Hora.Timestamp
+import Data.Time.Hora.Stamp
 import Data.Time.Hora.Future as F
-import Data.Time.Hora.Type.Time
+import Data.Time.Hora.Type
 
 
 
@@ -21,10 +21,6 @@ main = hspec $ do
             traceIO $ show diff3
             1 `shouldBe` 1
           it "DmyHmP" $ do
-            dmy2 <- now
+            dmy2 <- now::IO (DatePart Int)
             traceIO $ "TestDiffTime" ++ (show dmy2)
-            1 `shouldBe` 1
-          it "dmy_local" $ do
-            Tz tz1 l1 <- d'
-            traceIO $ show l1
             1 `shouldBe` 1